ref: move make_msg functions into message_helpers.py
